270 complete stripe payments set up payment intent recieved back

+++ client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -13,7 +13,7 @@ declare var Stripe;
   templateUrl: './checkout-payment.component.html',
   styleUrls: ['./checkout-payment.component.scss']
 })
-export class CheckoutPaymentComponent implements AfterViewInit {
+export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
   @Input() checkoutForm: FormGroup;
   @ViewChild('cardNumber', {static: true}) cardNumberElement: ElementRef;
   @ViewChild('cardExpiry', {static: true}) cardExpiryElement: ElementRef;
@@ -29,12 +29,8 @@ export class CheckoutPaymentComponent implements AfterViewInit {
     private basketService: BasketService,
     private checkoutService: CheckoutService,
     private toastr: ToastrService,
-    private router: Router
-    ) { }
+    private router: Router) { }
 
- 
-
-  // tslint:disable-next-line: typedef
   ngAfterViewInit() {
     this.stripe = Stripe('pk_test_51IhJzxLRFCWIs366rsuNIXdDIPxL82HvdGerSfjHBPcU21iel0YYGyw4jhnSfIr3mlqckOEyOcm62MFP9U6P0LEu00O819mKnI');
     const elements = this.stripe.elements();
@@ -49,51 +45,57 @@ export class CheckoutPaymentComponent implements AfterViewInit {
 
     this.cardCvc = elements.create('cardCvc');
     this.cardCvc.mount(this.cardCvcElement.nativeElement);
-    this.cardCvc.addEventListener('change', this.cardHandler);
-  }
-
- 
-
-  // tslint:disable-next-line: typedef
-  onChange({error}) {
-    if(error) {
-      this.cardErrors = error.message;
-    } else {
-      this.cardErrors = null;
-    }
-
-  }
-
-  // tslint:disable-next-line: typedef
-  submitOrder() {
-    // on successful basket additon routes to success page
-    const basket = this.basketService.getCurrentBasketValue();
-    const orderToCreate = this.getOrderToCreate(basket);
-    this.checkoutService.createOrder(orderToCreate).subscribe((order: IOrder) => {
-      this.toastr.success('Order created successfully');
-      this.basketService.deleteLocalBasket(basket.id);
-      const navigationExtras: NavigationExtras = { state: order };
-      this.router.navigate(['checkout/success'], navigationExtras );
-      console.log('the order is: ', order);
-    }, error => {
-      this.toastr.error(error.message);
-      console.log('submit order method in checkout payment error ', error);
-    });
-  }
-
-  // tslint:disable-next-line: typedef
-  private getOrderToCreate(basket: IBasket) {
-    return {
-      basketId: basket.id,
-      deliveryMethodId: + this.checkoutForm.get('deliveryForm'). get('deliveryMethod').value,
-      shipToAddress: this.checkoutForm.get('addressForm').value
-    };
+    this.cardNumber.addEventListener('change', this.cardHandler);
   }
 
   ngOnDestroy() {
     this.cardNumber.destroy();
     this.cardExpiry.destroy();
     this.cardCvc.destroy();
+  }
+
+  onChange({error}) {
+    if (error) {
+      this.cardErrors = error.message;
+    } else {
+      this.cardErrors = null;
+    }
+  }
+
+  submitOrder() {
+    const basket = this.basketService.getCurrentBasketValue();
+    const orderToCreate = this.getOrderToCreate(basket);
+    this.checkoutService.createOrder(orderToCreate).subscribe((order: IOrder) => {
+      this.toastr.success('Order created successfully');
+      this.stripe.confirmCardPayment(basket.clientSecret, {
+        payment_method: {
+          card: this.cardNumber,
+          billing_details: {
+            name: this.checkoutForm.get('paymentForm').get('nameOnCard').value
+          }
+        }
+      }).then(result => {
+        console.log(result);
+        if (result.paymentIntent) {
+          this.basketService.deleteLocalBasket(basket.id);
+          const navigationExtras: NavigationExtras = {state: order};
+          this.router.navigate(['checkout/success'], navigationExtras);
+        } else {
+          this.toastr.error(result.error.message);
+        }
+      });
+    }, error => {
+      this.toastr.error(error.message);
+      console.log(error);
+    });
+  }
+
+  private getOrderToCreate(basket: IBasket) {
+    return {
+      basketId: basket.id,
+      deliveryMethodId: +this.checkoutForm.get('deliveryForm').get('deliveryMethod').value,
+      shipToAddress: this.checkoutForm.get('addressForm').value
+    };
   }
 
 }
